requirements.txt
# Core dependencies
Flask>=2.0.1
Flask-SQLAlchemy>=2.5.1
Flask-Migrate>=3.1.0
SQLAlchemy>=1.4.23
python-dotenv>=0.19.0

# PDF Processing
pdfplumber>=0.7.4
PyPDF2>=2.10.5

# AI/ML
openai>=0.27.8
numpy>=1.26.0
pandas>=2.0.0

# Authentication
authlib>=1.0.1
requests>=2.26.0

# Database
alembic>=1.7.1

# Task Scheduling
APScheduler>=3.9.1

# Development
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2
